+++ crates/services/gas_price_service/src/v1/service.rs
@@ -279,9 +279,8 @@ where
             new_recorded_height = Some(BlockHeight::from(end));
         }
 
-        tracing::info!("Updating recorded height to {:?}", new_recorded_height);
         if let Some(recorded_height) = new_recorded_height {
-            tracing::info!("inner {:?}", recorded_height);
+            tracing::debug!("Updating recorded height to {:?}", recorded_height);
             storage_tx
                 .set_recorded_height(recorded_height)
                 .map_err(|err| anyhow!(err))?;