+++ crates/fuel-core/src/schema/node_info.rs
@@ -3,11 +3,18 @@ use super::scalars::{
     U64,
 };
 use crate::{
+    database::database_description::IndexationKind,
     fuel_core_graphql_api::{
         query_costs,
         Config as GraphQLConfig,
     },
-    graphql_api::api_service::TxPool,
+    graphql_api::{
+        api_service::TxPool,
+        database::{
+            IndexationFlags,
+            ReadDatabase,
+        },
+    },
 };
 use async_graphql::{
     Context,
@@ -23,6 +30,7 @@ pub struct NodeInfo {
     max_size: U64,
     max_depth: U64,
     node_version: String,
+    indexation: IndexationFlags,
 }
 
 #[Object]
@@ -53,6 +61,10 @@ impl NodeInfo {
 
     async fn node_version(&self) -> String {
         self.node_version.to_owned()
+    }
+
+    async fn indexation(&self) -> &IndexationFlags {
+        &self.indexation
     }
 
     #[graphql(complexity = "query_costs().storage_read + child_complexity")]
@@ -94,6 +106,8 @@ impl NodeQuery {
 
         const VERSION: &str = env!("CARGO_PKG_VERSION");
 
+        let db = ctx.data_unchecked::<ReadDatabase>();
+        let read_view = db.view()?;
         Ok(NodeInfo {
             utxo_validation: config.utxo_validation,
             vm_backtrace: config.vm_backtrace,
@@ -102,6 +116,7 @@ impl NodeQuery {
             max_size: (config.max_size as u64).into(),
             max_depth: (config.max_txpool_dependency_chain_length as u64).into(),
             node_version: VERSION.to_owned(),
+            indexation: read_view.indexation_flags,
         })
     }
 }
@@ -166,5 +181,23 @@ impl TxPoolStats {
     /// The total gas of the transactions in the pool
     async fn total_gas(&self) -> U64 {
         self.0.total_gas.into()
+    }
+}
+
+#[Object]
+impl IndexationFlags {
+    /// Is balances indexation enabled
+    async fn balances(&self) -> bool {
+        self.contains(&IndexationKind::Balances)
+    }
+
+    /// Is coins to spend indexation enabled
+    async fn coins_to_spend(&self) -> bool {
+        self.contains(&IndexationKind::CoinsToSpend)
+    }
+
+    /// Is asset metadata indexation enabled
+    async fn asset_metadata(&self) -> bool {
+        self.contains(&IndexationKind::AssetMetadata)
     }
 }